test: move missing paths tests to another module
