+++ src/parser/multi_module/tests/mod.rs
@@ -1,3 +1,4 @@
 mod duplication_path_error;
+mod missing_path;
 mod positive;
 mod utils;
